+++ src/HashTable.cpp
@@ -16,12 +16,12 @@ using namespace std;
 
 HashTable::HashTable(const Relation& relation,
                      uint32_t buckets,
-                     uint32_t (* const hashFunction)(uint32_t, int32_t)) :
+                     uint32_t (* const hashFunction)(uint32_t, uint64_t)) :
         buckets(buckets),
         numTuples(relation.getNumTuples()),
         hashFunction(hashFunction),
-        histogram(new uint32_t[buckets] { }),
-        pSum(new uint32_t[buckets] { }),
+        histogram(new uint64_t[buckets] { }),
+        pSum(new uint64_t[buckets] { }),
         orderedTuples(new Tuple[relation.getNumTuples()] { }) {
     ConsoleOutput consoleOutput("HashTable");
     if (this->buckets == 0) {
@@ -39,7 +39,7 @@ HashTable::HashTable(const Relation& relation,
                "Splitting " << numTuples << " tuples to " << (this->buckets) << " buckets");
 
     CO_IFDEBUG(consoleOutput, "Generating histogram");
-    for (uint32_t i = 0; i < numTuples; ++i) {
+    for (uint64_t i = 0; i < numTuples; ++i) {
         CO_IFDEBUG(consoleOutput, "Processing tuple " << i);
         const Tuple& toCheck = relation.getTuple(i);
         CO_IFDEBUG(consoleOutput, i << ":" << toCheck);
@@ -60,7 +60,7 @@ HashTable::HashTable(const Relation& relation,
 
     CO_IFDEBUG(consoleOutput, "Generating pSum");
     {
-        uint32_t prevSum = 0;
+        uint64_t prevSum = 0;
         for (uint32_t i = 0; i < buckets; ++i) {
             CO_IFDEBUG(consoleOutput,
                        "\t[bucket=" << i << ", pSum=" << prevSum << "]");
@@ -72,7 +72,7 @@ HashTable::HashTable(const Relation& relation,
     CO_IFDEBUG(consoleOutput, "pSum generated");
 
     CO_IFDEBUG(consoleOutput, "Copying Tuples");
-    for (uint32_t i = 0; i < numTuples; ++i) {
+    for (uint64_t i = 0; i < numTuples; ++i) {
         CO_IFDEBUG(consoleOutput, "Processing tuple " << i);
         const Tuple& toCheck = relation.getTuple(i);
         CO_IFDEBUG(consoleOutput, i << ":" << toCheck);
@@ -90,11 +90,11 @@ HashTable::HashTable(const HashTable & toCopy) :
         buckets(toCopy.buckets),
         numTuples(toCopy.numTuples),
         hashFunction(toCopy.hashFunction),
-        histogram(new uint32_t[buckets]),
-        pSum(new uint32_t[buckets]),
+        histogram(new uint64_t[buckets]),
+        pSum(new uint64_t[buckets]),
         orderedTuples(new Tuple[toCopy.numTuples]) {
-    memcpy(histogram, toCopy.histogram, buckets * sizeof(uint32_t));
-    memcpy(pSum, toCopy.pSum, buckets * sizeof(uint32_t));
+    memcpy(histogram, toCopy.histogram, buckets * sizeof(uint64_t));
+    memcpy(pSum, toCopy.pSum, buckets * sizeof(uint64_t));
     memcpy(orderedTuples, toCopy.orderedTuples, numTuples * sizeof(Tuple));
 }
 
@@ -108,7 +108,7 @@ uint32_t HashTable::getBuckets() const {
     return buckets;
 }
 
-uint32_t HashTable::getTuplesInBucket(uint32_t bucket) const {
+uint64_t HashTable::getTuplesInBucket(uint32_t bucket) const {
     if (bucket >= buckets) {
         throw runtime_error("bucket out of bounds [bucket="
                             + to_string(bucket)
@@ -132,7 +132,7 @@ const Tuple * const HashTable::getBucket(uint32_t bucket) const {
     return orderedTuples + pSum[bucket];
 }
 
-const Tuple& HashTable::getTuple(uint32_t bucket, uint32_t index) const {
+const Tuple& HashTable::getTuple(uint32_t bucket, uint64_t index) const {
     if (bucket >= buckets) {
         throw runtime_error("bucket out of bounds [bucket="
                             + to_string(bucket)
@@ -177,10 +177,10 @@ std::ostream& operator<<(std::ostream& os, const HashTable& toPrint) {
     os << "], orderedTuples=[";
     for (uint32_t i = 0; i < toPrint.buckets; ++i) {
         os << "\n\tBucket #" << i;
-        uint32_t bucketStart = toPrint.pSum[i];
-        uint32_t numInBucket = toPrint.histogram[i];
-        for (uint32_t j = 0; j < numInBucket; ++j) {
-            uint32_t pos = bucketStart + j;
+        uint64_t bucketStart = toPrint.pSum[i];
+        uint64_t numInBucket = toPrint.histogram[i];
+        for (uint64_t j = 0; j < numInBucket; ++j) {
+            uint64_t pos = bucketStart + j;
             os << "\n\t\t<"
                << pos
                << ","
